requirements.txt
matplotlib
numpy
plotly
scipy
seaborn
umap-learn
